% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{billy}
\alias{billy}
\title{Billy Joel.}
\format{
double array containing the pixel values of Billy Joels image of the LFWcrop data set.
}
\source{
\url{https://towardsdatascience.com/eigenfaces-recovering-humans-from-ghosts-17606c328184}
}
\usage{
billy
}
\description{
Billy Joels image of the LFWcrop data set. Please refer to \code{?images} for more information.
}
\keyword{datasets}
