% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{change_max_eigenfaces}
\alias{change_max_eigenfaces}
\title{Adjust Number of Eigenfaces.}
\usage{
change_max_eigenfaces(dataset, max_eigenfaces = 0L)
}
\arguments{
\item{dataset}{list; List returned by load_dataset() with computed eigenfaces. (required)}

\item{max_eigenfaces}{integer; Number of eigenfaces that will be computed. (optional)}
}
\value{
A list containing the eigenfaces and other information (\code{?dataset} for more information) with new number \code{max_eigenfaces} of eigenfaces.
}
\description{
Change the number of eigenfaces originally specified as \code{max_eigenfaces} in \code{load_dataset}.
}
\examples{
change_max_eigenfaces(dataset, 42)
}
