% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Eigenfaces of LFWcrop data set.}
\format{
A list with 5 entires:
\describe{
\item{eigenfaces}{n = max_eigenfaces eigenfaces/eigenvectors}
\item{mean}{the mean image of the data set}
\item{image_size}{resolution of the images}
\item{std_values}{normalized pixel values (original - mean)}
\item{image_coef}{coefficients to reconstruct original face image from eigenfaces}
}
}
\usage{
dataset
}
\description{
50 eigenfaces of 100 LSWcrop images of celebrities. Please refer to \code{?images} for more information.
}
\keyword{datasets}
