% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{load_dataset}
\alias{load_dataset}
\title{Load Training Data and Compute Eigenfaces.}
\usage{
load_dataset(path, pattern = NULL, max_images = 0L, max_eigenfaces = 0L)
}
\arguments{
\item{path}{character; Specify the path to the data set. (required)}

\item{pattern}{character; Regular expression used to filter images of the data set. Only image files matching the RegEx will be used. (optional)}

\item{max_images}{integer; Maximum number of images used for eigenface computation. (optional)}

\item{max_eigenfaces}{integer; Number of eigenfaces that will be computed. (optional)}
}
\value{
A list containing the eigenfaces and other information (\code{?dataset} for more information).
}
\description{
The specified data set is loaded and face images are converted to face vectors. The mean face is subtracted from all face vectors. Subsequently, the covariance matrix is computed and eigenvalues and corresponding eigenvectors (the eigenfaces) are determined.
}
\examples{
\dontrun{
my_dataset <- REigenfaces::load_dataset("C:/Users/me/data/lfwcrop_grey/faces",
                                        max_images=13000,
                                        pattern="(0001)|(0002)|(0003)|(0004)",
                                        max_eigenfaces=100L)
}
}
