% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{load_images}
\alias{load_images}
\title{Load Images.}
\usage{
load_images(path, pattern = NULL, max_images = 0L)
}
\arguments{
\item{path}{character; Specify the path to the image(s). (required)}

\item{pattern}{character; Regular expression used to filter images of the data set. Only image files matching the RegEx will be used. (optional)}

\item{max_images}{integer, Number of images that will be loaded. (optional)}
}
\value{
A double vector containing the loaded image(s).
}
\description{
Load (new) image(s) to search similar faces in the training data set.
}
\examples{
\dontrun{
images <- load_images("C:/Users/me/data/images", pattern="(0001)|(0002)|(0003)|(0004)",
                       max_images=1L)
}
}
