% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{load_pgm_images}
\alias{load_pgm_images}
\title{Load pgm Images.}
\usage{
load_pgm_images(path, pattern = NULL, max_images = 0L)
}
\arguments{
\item{path}{character; Specify the path to the image(s). (required)}

\item{pattern}{character; Regular expression used to filter images of the data set. Only image files matching the RegEx will be used. (optional)}

\item{max_images}{integer, Number of images that will be loaded. (optional)}
}
\value{
A double vector containing the loaded image(s).
}
\description{
Load (new) image(s) to compute eigenfaces or search similar faces in the training data set. Only images in the pgm image format (greymap) are supported.
}
\examples{
\dontrun{
images <- load_pgm_images("C:/Users/me/data/lfwcrop_grey/faces",
                          pattern="(0001)|(0002)|(0003)|(0004)",
                          max_images=1000L)
}
}
