% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{reconstruct_dataset_images}
\alias{reconstruct_dataset_images}
\title{Reconstruct Faces Using Eigenfaces.}
\usage{
reconstruct_dataset_images(dataset, indices)
}
\arguments{
\item{dataset}{list; List returned by load_dataset() with computed eigenfaces. (required)}

\item{indices}{integer; Indices of the original images that will be reconstructed. (required)}
}
\description{
Reconstruct the original training data faces by linear combinations of eigenfaces.
}
\examples{
reconstruct_dataset_images(dataset, 1:16)
}
