% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REigenfaces-package.R
\docType{package}
\name{REigenfaces-package}
\alias{REigenfaces}
\alias{REigenfaces-package}
\title{REigenfaces: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
mypkg is an awesome package.
It contains awesome functions, like \code{\link{add}}.
}
\references{
\url{https://de.wikipedia.org/wiki/Eigengesichter}
}
\author{
\strong{Maintainer}: Max Edinger \email{iu220@stud.uni-heidelberg.de} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
