% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Shiny App Example}
\usage{
runShiny()
}
\description{
Shiny App presenting the package functionality. Select one of the 100 images in the top left to see the eigenface representation, similar faces and the contribution of each eigenface to the representation.
}
\examples{
\dontrun{
runShiny()
}
}
