% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{show_images}
\alias{show_images}
\title{Show Images.}
\usage{
show_images(images, size)
}
\arguments{
\item{images}{double; Vector containing the images that will be displayed. (required)}

\item{size}{integer; Resolution of the image(s). (required)}
}
\description{
Show images that have been loaded via \code{load_pgm_images} or eigenfaces that have been computed using \code{load_dataset} or \code{most_important_eigenfaces} or \code{reconstructed_dataset_images}.
}
\examples{
show_images(billy, dataset$image_size)
}
