% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{show_most_important_eigenfaces}
\alias{show_most_important_eigenfaces}
\title{Display Most Important Eigenfaces.}
\usage{
show_most_important_eigenfaces(dataset, max_count = 1)
}
\arguments{
\item{dataset}{list; List returned by load_dataset() with computed eigenfaces. (required)}

\item{max_count}{integer; Number of eigenfaces that will be displayed. (optional)}
}
\description{
The \code{max_count} most important eigenfaces (eigenvectors corresponding to the \code{max_count} eigenvalues with highest absolute value) are displayed.
}
\examples{
show_most_important_eigenfaces(dataset, 16)
}
