% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{show_similar_faces}
\alias{show_similar_faces}
\title{Display Similar Faces.}
\usage{
show_similar_faces(dataset, image, max_count = 1)
}
\arguments{
\item{dataset}{list; List returned by load_dataset() with computed eigenfaces. (required)}

\item{image}{numeric; Image, for which similar faces are determined. (required)}

\item{max_count}{integer; Number of similar faces that will be displayed. (optional)}
}
\description{
\code{max_count} similar faces of the training data set are displayed.
}
\examples{
show_similar_faces(dataset, dataset$images[1], 16)
}
