% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{load_dataset}
\alias{load_dataset}
\title{Load Training Data and Compute Eigenfaces.}
\usage{
load_dataset(images, max_eigenfaces = 0L, standardized = TRUE)
}
\arguments{
\item{images}{double array; Images for which eigenfaces will be computed. (required)}

\item{max_eigenfaces}{integer; Number of eigenfaces that will be computed. (optional)}

\item{standardized}{logical; Specify whether or not images should be standerdized (subtract mean image from every image). (optional)}
}
\value{
A list containing the eigenfaces and other information (\code{?dataset} for more information).
}
\description{
The specified data set is loaded (images that have been loaded with \code{load_pgm_images} before) and face images are converted to face vectors. The mean face is subtracted from all face vectors. Subsequently, the covariance matrix is computed and eigenvalues and corresponding eigenvectors (the eigenfaces) are determined.
}
\examples{
\dontrun{
my_dataset <- load_dataset(my_dataset_images, max_eigenfaces=100L, standardized=TRUE)
}
}
